requirements.txt
 # Core dependencies
  fastapi==0.104.1
  uvicorn[standard]==0.24.0
  python-multipart==0.0.6

  # AWS dependencies
  boto3==1.34.0
  botocore==1.34.0

  # OpenAI and embeddings
  openai==1.3.0
  tiktoken==0.5.1

  # Vector database
  pinecone-client==2.2.4

  # Document processing
  pypdf2==3.0.1
  python-docx==1.1.0

  # Database
  psycopg2-binary==2.9.9
  sqlalchemy==2.0.23

  # Environment management
  python-dotenv==1.0.0
